Updating file names and "parser" related functions
