+++ scripts/plugins/config/cli.py
